Settings dialog rethink
Settings are now organised into GENERAL and FAKE DATA.
Randomize on boot checkbox disables the randomisable data inputs.

Added a checkbox in app settings to use global settings. The detection 
code is a new function in PrivacyManager which checks to see if serial#
returns null for that uid. The result is functionality similar to the 
old use of the Clear button. The checkbox also disables the fake data 
inputs.

Two problems remain:
1. strings need to be added for all languages.
2. textedits at the bottom of the screen get covered by the keyboard.