Update README.md
Notes: when a sentence contains the agent and object of a transitive verb, clarity and directness of expression prefer the active voice over the passive. Similarly (though less seriously), the modal verb "will" is deadwood because, in this context, there's no difference between what stating what XPrivacy does and what XPrivacy will do.