edited installation, upgrading, and usage sections
Note: when an English sentence contains the agent and object for a verb, clarity and directness of expression require the active voice over the passive. Also, the modal verb "will" is frequently deadwood. In the present context, there's no difference in meaning between "XPrivacy will <do sth.>" and "XPrivacy <does sth.>."