Anchor adjustments
The Github markdown system allows anchors on the same line as headings, so I moved the FAQ anchor onto the same line as the header.

I also added a *news* anchor for the *In The Media* section.  The *news* heading was confusing because the user may have expected to read news from the developer regarding XPrivacy in that section (which was not the case).  Although it will now work either way, you may want to change the *News* link in XDA post 1 to read *In The Media*.  At the same time, you can add a *Screenshots* link in XDA post 1, if you desire.